Modify fmt and folly compile logic in build.sh

+++ cpp-omni/build.sh
@@ -19,20 +19,20 @@ set -e
 
 install_fmt() {
   echo "Start build fmt.so"
-  local fmt_so_core="${OMNI_HOME}/lib/libfmt.so.10"
   local fmt_tag="10.1.1"
   local fmt_repo="https://gitee.com/mirrors/fmt.git"
   local open_source_dir="open_source"
   local workspace=$(readlink -f $(dirname "$BASH_SOURCE"))
   local fmt_source_dir="${workspace}/${open_source_dir}/fmt"
-  local fmt_build_dir="${fmt_source_dir}/build"
+  local fmt_default_home="/usr/local"
 
-  if [ ! -f "${fmt_so_core}" ]; then
-    echo ">>>>> libfmt.so.10 not found in ${OMNI_HOME}/lib, start to clone fmt-${fmt_tag} source code and build..."
+  if [ -n "$FMT_HOME" ] && [ -d "$FMT_HOME" ]; then
+    echo ">>>>> FMT_HOME=$FMT_HOME exists, skip fmt build process."
+  else
+    echo ">>>>> FMT_HOME not set, start to clone fmt-${fmt_tag} source code and build..."
     rm -rf ${fmt_source_dir} && mkdir -p ${fmt_source_dir}
     git clone --branch ${fmt_tag} --depth=1 ${fmt_repo} ${fmt_source_dir}
     cd ${fmt_source_dir} && mkdir -p build && cd build
-    # Cmake build with your specified params
     cmake .. \
     -DCMAKE_BUILD_TYPE=Release \
     -DFMT_TEST=OFF \
@@ -41,13 +41,9 @@ install_fmt() {
     -DBUILD_SHARED_LIBS=ON
     make -j$(nproc)
     sudo make install
-    # Copy all generated fmt so files to OMNI_HOME/lib directly
-    sudo cp -f libfmt.so* ${OMNI_HOME}/lib/
-    echo ">>>>> All fmt shared libraries copied to ${OMNI_HOME}/lib successfully."
-    # Back to workspace
+    export FMT_HOME=${fmt_default_home}
+    echo ">>>>> Set FMT_HOME=$FMT_HOME automatically after fmt install."
     cd ../../../../
-  else
-    echo ">>>>> libfmt.so.10 already exists in ${OMNI_HOME}/lib, skip fmt build process."
   fi
 }
 
@@ -58,15 +54,22 @@ install_folly() {
   local open_source_dir="open_source"
   local workspace=$(readlink -f $(dirname "$BASH_SOURCE"))
   local folly_source_dir="${workspace}/${open_source_dir}/folly"
-  echo ">>>>> Start to clone folly-${folly_tag} source code and build..."
-  rm -rf ${folly_source_dir} && mkdir -p ${folly_source_dir}
-  git clone --branch ${folly_tag} --depth=1 ${folly_repo} ${folly_source_dir}
-  cd ${folly_source_dir} && mkdir -p build && cd build
-  cmake .. -DBUILD_TESTS=OFF -DFOLLY_HAVE_INT128_T=ON
-  make -j$(nproc)
-  sudo make install
-  echo ">>>>> folly-${folly_tag} build and install completed successfully."
-  cd ../../../../
+  local folly_default_home="/usr/local"
+
+  if [ -n "$FOLLY_HOME" ] && [ -d "$FOLLY_HOME" ]; then
+    echo ">>>>> FOLLY_HOME=$FOLLY_HOME exists, skip folly build process."
+  else
+    echo ">>>>> FOLLY_HOME not set, start to clone folly-${folly_tag} source code and build..."
+    rm -rf ${folly_source_dir} && mkdir -p ${folly_source_dir}
+    git clone --branch ${folly_tag} --depth=1 ${folly_repo} ${folly_source_dir}
+    cd ${folly_source_dir} && mkdir -p build && cd build
+    cmake .. -DBUILD_TESTS=OFF -DFOLLY_HAVE_INT128_T=ON
+    make -j$(nproc)
+    sudo make install
+    export FOLLY_HOME=${folly_default_home}
+    echo ">>>>> Set FOLLY_HOME=$FOLLY_HOME automatically after folly install."
+    cd ../../../../
+  fi
 }
 
 if [ -z "$OMNI_HOME" ]; then
